Proposed changes
Add a new function: floor_to.
Also added in the NAMESPACE file.
Version update.
In the AICc, dfgam, flexOR, plot.OR and predict.OR functions added the description tag.
Proposed change to the dfgam function.
Change in the plot.OR function - added the col.area argument + some proposed changes.
In the predict.OR function - change in the variable name (from pred to ref) and change in the return variable. (possibility of having to go back).